Add and Clear buttons now work
Added the button code into the addSalesRecordGUI file
