+++ src/php/GUIwrapper.java
@@ -4,11 +4,12 @@ import java.awt.*;
 import javax.swing.*;
 
 public class GUIwrapper {
-	public static JTabbedPane TabbedPane() {
+	public JTabbedPane TabbedPane() {
 		JTabbedPane tabbedPane = new JTabbedPane();
 		
 		JPanel addSalesGUI = new JPanel();
-		AddSalesRecordGUI.AddSalesContent(addSalesGUI);
+		AddSalesRecordGUI addSales = new AddSalesRecordGUI();
+		addSales.AddSalesContent(addSalesGUI);
 		tabbedPane.addTab("Add Sales", null, addSalesGUI, null);
 		return tabbedPane;
 	}
