+++ src/php/WindowForm.java
@@ -1,7 +1,6 @@
 package php;
 
 import java.awt.*;
-import java.awt.event.*;
 import javax.swing.*;
 
 public class WindowForm {
@@ -30,29 +29,9 @@ public class WindowForm {
 		frame = new JFrame("People Health Pharmacy");
 		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		//frame.setPreferredSize(new Dimension(800,600));
-		
-		frame.add(GUIwrapper.TabbedPane());
-		
-		/*
-		// Add Button
-		
-		
-		addButton.setBounds(99, 158, 69, 22);
-		frame.getContentPane().add(addButton);
-		
-		// Clear button
-		
-		clearButton.addActionListener(new ActionListener() {
-			public void actionPerformed(ActionEvent e) {
-				itemTextField.setText("");
-				quantityTextField.setText("");
-				priceTextField.setText("");	
-			}
-		});
-		clearButton.setBounds(210, 158, 69, 22);
-		frame.getContentPane().add(clearButton);
-		
-		*/
+		GUIwrapper gui = new GUIwrapper();
+		frame.add(gui.TabbedPane());
+
 		frame.pack();
 		frame.setLocationRelativeTo(null); // must go after .pack() to work
 	}
